Create Home Router and remove topnumbers and right numbers from landing router
